+++ src/main/scala/it/unipi/di/acubelab/wikipediarelatedness/wikipedia/relatedness/MilneWittenRelatedness.scala
@@ -13,17 +13,18 @@ class MilneWittenRelatedness(options: MilneWittenOptions) extends Relatedness {
   /**
     * Paper: https://www.aaai.org/Papers/Workshops/2008/WS-08-15/WS08-15-005.pdf
     */
-  def computeRelatedness(srcWikiID: Int, dstWikiID: Int) : Double = {
+  def computeRelatedness(srcWikiID: Int, dstWikiID: Int) : Float = {
     val sizeA = graph.outdegree(srcWikiID)
     val sizeB = graph.outdegree(dstWikiID)
 
     val intersection = graph.linkIntersection(srcWikiID, dstWikiID)
 
-    if (intersection == 0) return 0.0
+    if (intersection == 0) return 0.0f
 
     val rel = (math.log(sizeA max sizeB) - math.log(intersection) ) /
                 (math.log(W) - math.log(sizeA min sizeB))
-    val normRel = 1 - ((rel max 0.0) min 1.0)
+
+    val normRel = 1 - ((rel.toFloat max 0.0f) min 1.0f)
 
     normRel
   }
